Entrada de dados em c#

+++ DevSuperiorLinguagemEmCs/ProcessamentoDeDados/ProcessamentoDeDadosEmC#.cs
@@ -16,7 +16,7 @@ class ProcessamentoDeDadosEmC_ {
         b1 = 6.0;
         b2 = 8.0;
         h = 5.0;
-        area = (b1 + b2) / 2;
+        area = (double) (b1 + b2) / 2;
 
         Console.WriteLine(area.ToString("F2", CI));
     }
